switching to pypyroject.toml
