+++ .github/workflows/ci.yml
@@ -50,9 +50,6 @@ jobs:
       - name: make history
         run: python3 ./changes/make_history.py
 
-      - name: check dist
-        run: make check-dist
-
       - name: install node for pyright
         uses: actions/setup-node@v3
         with:
@@ -80,7 +77,7 @@ jobs:
           docs-build
           ${{ runner.os }}
           ${{ env.pythonLocation }}
-          ${{ hashFiles('setup.py') }}
+          ${{ hashFiles('pyproject.toml') }}
           ${{ hashFiles('requirements.txt') }}
           ${{ hashFiles('docs/requirements.txt') }}
 
@@ -104,6 +101,11 @@ jobs:
       matrix:
         os: [ubuntu, macos, windows]
         python-version: ['3.7', '3.8', '3.9', '3.10', '3.11.0-rc.1']
+        exclude:  # these are slow, we should remove once 3.11 is released
+          - os: 'macos'
+            python-version: '3.11.0-rc.1'
+          - os: 'windows'
+            python-version: '3.11.0-rc.1'
 
     env:
       PYTHON: ${{ matrix.python-version }}
@@ -127,7 +129,7 @@ jobs:
           test
           ${{ runner.os }}
           ${{ env.pythonLocation }}
-          ${{ hashFiles('setup.py') }}
+          ${{ hashFiles('pyproject.toml') }}
           ${{ hashFiles('requirements.txt') }}
           ${{ hashFiles('tests/requirements-testing.txt') }}
 
@@ -247,68 +249,10 @@ jobs:
     - name: test
       run: make test-fastapi
 
-  build:
-    name: build py3.${{ matrix.python-version }} on ${{ matrix.platform || matrix.os }}
-    needs: [lint, test, test-old-mypy, test-fastapi]
-    if: "success() && (startsWith(github.ref, 'refs/tags/') || github.ref == 'refs/heads/main')"
-    strategy:
-      fail-fast: false
-      matrix:
-        os: [ubuntu , macos , windows]
-        python-version: ['7', '8', '9', '10', '11']
-        include:
-        - os: ubuntu
-          platform: linux
-        - os: windows
-          ls: dir
-
-    runs-on: ${{ matrix.os }}-latest
-    steps:
-    - uses: actions/checkout@v3
-
-    - name: set up python
-      uses: actions/setup-python@v4
-      with:
-        python-version: '3.8'
-
-    - name: install
-      run: pip install -U twine setuptools wheel cibuildwheel
-
-    - name: build sdist
-      if: matrix.os == 'ubuntu' && matrix.python-version == '9'
-      run: python setup.py sdist bdist_wheel
-
-    - name: build ${{ matrix.platform || matrix.os }} binaries
-      run: cibuildwheel --output-dir dist
-      env:
-        PIP: 'pip'
-        CIBW_BUILD: 'cp3${{ matrix.python-version }}-*'
-        CIBW_SKIP: '*-win32'
-        CIBW_PLATFORM: '${{ matrix.platform || matrix.os }}'
-        CIBW_TEST_REQUIRES: 'pytest==6.2.5 pytest-mock==3.6.1'
-        CIBW_TEST_COMMAND: 'pytest {project}/tests'
-        CIBW_MANYLINUX_X86_64_IMAGE: 'manylinux2014'
-        CIBW_MANYLINUX_I686_IMAGE: 'manylinux2014'
-        CIBW_ARCHS_MACOS: 'x86_64 arm64'
-        CIBW_TEST_SKIP: '*-macosx_arm64'  # see https://cibuildwheel.readthedocs.io/en/stable/faq/#universal2
-
-    # TODO build windows 32bit binaries
-
-    - name: list dist files
-      run: |
-        ${{ matrix.ls || 'ls -lh' }} dist/
-        twine check dist/*
-
-    - name: Store dist artifacts
-      uses: actions/upload-artifact@v3
-      with:
-        name: pypi_files
-        path: dist
-
   deploy:
     name: Deploy
-    needs: build
-    if: "success() && startsWith(github.ref, 'refs/tags/')"
+    needs: [lint, test, test-old-mypy, test-fastapi]
+#    if: "success() && startsWith(github.ref, 'refs/tags/')"
     runs-on: ubuntu-latest
 
     steps:
@@ -319,12 +263,6 @@ jobs:
         with:
           python-version: '3.10'
 
-      - name: get dist artifacts
-        uses: actions/download-artifact@v3
-        with:
-          name: pypi_files
-          path: dist
-
       - name: get docs
         uses: actions/download-artifact@v3
         with:
@@ -332,25 +270,26 @@ jobs:
           path: site
 
       - name: install
-        run: pip install -U twine packaging
-
-      - name: twine check
-        run: |
-          twine check dist/*
-          ls -lh dist
+        run: pip install -U twine build packaging
 
       - name: check tag
         id: check-tag
         run: ./tests/check_tag.py
 
-      - name: upload to pypi
-        run: twine upload dist/*
-        env:
-          TWINE_USERNAME: __token__
-          TWINE_PASSWORD: ${{ secrets.pypi_token }}
+      - name: build
+        run: python -m build
 
-      - name: publish docs
-        if: '!fromJSON(steps.check-tag.outputs.IS_PRERELEASE)'
-        run: make publish-docs
-        env:
-          NETLIFY: ${{ secrets.netlify_token }}
+      - run: twine check dist/*
+
+#      - name: upload to pypi
+#        run: twine upload dist/*
+#        env:
+#          TWINE_USERNAME: __token__
+#          TWINE_PASSWORD: ${{ secrets.pypi_token }}
+#          TWINE_PASSWORD: ${{ secrets.pypi_token }}
+#
+#      - name: publish docs
+#        if: '!fromJSON(steps.check-tag.outputs.IS_PRERELEASE)'
+#        run: make publish-docs
+#        env:
+#          NETLIFY: ${{ secrets.netlify_token }}
