+++ .pre-commit-config.yaml
@@ -1,10 +1,12 @@
 repos:
 - repo: https://github.com/pre-commit/pre-commit-hooks
-  rev: v4.0.1
+  rev: v4.3.0
   hooks:
   - id: check-yaml
     args: ['--unsafe']
+  - id: check-toml
   - id: end-of-file-fixer
+  - id: trailing-whitespace
 
 - repo: local
   hooks:
@@ -13,13 +15,16 @@ repos:
     entry: make lint
     types: [python]
     language: system
+    pass_filenames: false
   - id: mypy
     name: Mypy
     entry: make mypy
     types: [python]
     language: system
+    pass_filenames: false
   - id: pyupgrade
     name: Pyupgrade
     entry: make pyupgrade
     types: [python]
     language: system
+    pass_filenames: false
