--- MANIFEST.in
@@ -1,6 +1,0 @@
-include LICENSE
-include README.md
-include HISTORY.md
-graft tests
-global-exclude __pycache__
-global-exclude *.py[cod]
