+++ requirements.txt
@@ -2,6 +2,5 @@
 
 devtools==0.9.0
 email-validator==1.2.1
-dataclasses==0.6; python_version < '3.7'
 typing-extensions==4.3.0
 python-dotenv==0.20.0
