--- setup.cfg
@@ -1,92 +1,0 @@
-[tool:pytest]
-testpaths = tests
-addopts = -p no:hypothesispytest
-filterwarnings =
-    error
-    ignore::DeprecationWarning:distutils
-    ignore::DeprecationWarning:Cython
-    # for Python 3.10+: mypy still relies on distutils on windows. We hence ignore those warnings
-    ignore:The distutils package is deprecated and slated for removal in Python 3.12:DeprecationWarning
-    ignore:The distutils.sysconfig module is deprecated, use sysconfig instead:DeprecationWarning
-    # for Python 3.11
-    ignore:path is deprecated.*:DeprecationWarning:certifi
-    ignore:module 'sre_constants' is deprecated:DeprecationWarning:pkg_resources
-
-[flake8]
-max-line-length = 120
-max-complexity = 14
-inline-quotes = single
-multiline-quotes = double
-ignore = E203, W503
-per-file-ignores =
-    docs/examples/schema_unenforced_constraints.py: F811
-    docs/examples/validation_decorator_async.py: E402
-    docs/examples/types_constrained.py: F722
-
-[coverage:run]
-source = pydantic
-branch = True
-context = ${CONTEXT}
-
-[coverage:report]
-precision = 2
-exclude_lines =
-    pragma: no cover
-    raise NotImplementedError
-    raise NotImplemented
-    if TYPE_CHECKING:
-    @overload
-
-[coverage:paths]
-source =
-    pydantic/
-    /Users/runner/work/pydantic/pydantic/pydantic/
-    D:\a\pydantic\pydantic\pydantic
-
-[isort]
-line_length=120
-known_first_party=pydantic
-multi_line_output=3
-include_trailing_comma=True
-force_grid_wrap=0
-combine_as_imports=True
-
-[mypy]
-python_version = 3.9
-show_error_codes = True
-follow_imports = silent
-strict_optional = True
-warn_redundant_casts = True
-warn_unused_ignores = True
-disallow_any_generics = True
-check_untyped_defs = True
-no_implicit_reexport = True
-warn_unused_configs = True
-disallow_subclassing_any = True
-disallow_incomplete_defs = True
-disallow_untyped_decorators = True
-disallow_untyped_calls = True
-
-# for strict mypy: (this is the tricky one :-))
-disallow_untyped_defs = True
-
-# remaining arguments from `mypy --strict` which cause errors
-;no_implicit_optional = True
-;warn_return_any = True
-
-[mypy-email_validator]
-ignore_missing_imports = true
-
-[mypy-dotenv]
-ignore_missing_imports = true
-
-[mypy-toml]
-ignore_missing_imports = true
-
-# ansi2html and devtools are required to avoid the need to install these packages when running linting,
-# they're used in the docs build script
-[mypy-ansi2html]
-ignore_missing_imports = true
-
-[mypy-devtools]
-ignore_missing_imports = true
